+++ src/StardewModdingAPI/Framework/ContentHelper.cs
@@ -45,11 +45,11 @@ namespace StardewModdingAPI.Framework
 
         /// <summary>Load content from the game folder or mod folder (if not already cached), and return it. When loading a <c>.png</c> file, this must be called outside the game's draw loop.</summary>
         /// <typeparam name="T">The expected data type. The main supported types are <see cref="Texture2D"/> and dictionaries; other types may be supported by the game's content pipeline.</typeparam>
-        /// <param name="key">The asset key to fetch (if the <paramref name="source"/> is <see cref="ContentSource.GameContent"/>), or the local path to an XNB file relative to the mod folder.</param>
+        /// <param name="key">The asset key to fetch (if the <paramref name="source"/> is <see cref="ContentSource.GameContent"/>), or the local path to a content file relative to the mod folder.</param>
         /// <param name="source">Where to search for a matching content asset.</param>
         /// <exception cref="ArgumentException">The <paramref name="key"/> is empty or contains invalid characters.</exception>
         /// <exception cref="ContentLoadException">The content asset couldn't be loaded (e.g. because it doesn't exist).</exception>
-        public T Load<T>(string key, ContentSource source)
+        public T Load<T>(string key, ContentSource source = ContentSource.ModFolder)
         {
             this.AssertValidAssetKeyFormat(key);
             try
@@ -57,25 +57,22 @@ namespace StardewModdingAPI.Framework
                 switch (source)
                 {
                     case ContentSource.GameContent:
-                        return this.ContentManager.Load<T>(this.StripXnbExtension(key));
+                        return this.ContentManager.Load<T>(key);
 
                     case ContentSource.ModFolder:
-                        // find content file
-                        key = this.ContentManager.NormalisePathSeparators(key);
-                        FileInfo file = new FileInfo(Path.Combine(this.ModFolderPath, key));
-                        if (!file.Exists && file.Extension == "")
-                            file = new FileInfo(Path.Combine(this.ModFolderPath, key + ".xnb"));
+                        // get file
+                        FileInfo file = this.GetModFile(key);
                         if (!file.Exists)
                             throw new ContentLoadException($"There is no file at path '{file.FullName}'.");
 
-                        // get underlying asset key
-                        string actualKey = this.GetActualAssetKey(key, source);
+                        // get asset path
+                        string assetPath = this.GetModAssetPath(key, file.FullName);
 
                         // load content
                         switch (file.Extension.ToLower())
                         {
                             case ".xnb":
-                                return this.ContentManager.Load<T>(actualKey);
+                                return this.ContentManager.Load<T>(assetPath);
 
                             case ".png":
                                 // validate
@@ -83,15 +80,15 @@ namespace StardewModdingAPI.Framework
                                     throw new ContentLoadException($"Can't read file with extension '{file.Extension}' as type '{typeof(T)}'; must be type '{typeof(Texture2D)}'.");
 
                                 // try cache
-                                if (this.ContentManager.IsLoaded(actualKey))
-                                    return this.ContentManager.Load<T>(actualKey);
+                                if (this.ContentManager.IsLoaded(assetPath))
+                                    return this.ContentManager.Load<T>(assetPath);
 
                                 // fetch & cache
                                 using (FileStream stream = File.OpenRead(file.FullName))
                                 {
                                     Texture2D texture = Texture2D.FromStream(Game1.graphics.GraphicsDevice, stream);
                                     texture = this.PremultiplyTransparency(texture);
-                                    this.ContentManager.Inject(actualKey, texture);
+                                    this.ContentManager.Inject(assetPath, texture);
                                     return (T)(object)texture;
                                 }
 
@@ -110,19 +107,19 @@ namespace StardewModdingAPI.Framework
         }
 
         /// <summary>Get the underlying key in the game's content cache for an asset. This can be used to load custom map tilesheets, but should be avoided when you can use the content API instead. This does not validate whether the asset exists.</summary>
-        /// <param name="key">The asset key to fetch (if the <paramref name="source"/> is <see cref="ContentSource.GameContent"/>), or the local path to an XNB file relative to the mod folder.</param>
+        /// <param name="key">The asset key to fetch (if the <paramref name="source"/> is <see cref="ContentSource.GameContent"/>), or the local path to a content file relative to the mod folder.</param>
         /// <param name="source">Where to search for a matching content asset.</param>
         /// <exception cref="ArgumentException">The <paramref name="key"/> is empty or contains invalid characters.</exception>
-        public string GetActualAssetKey(string key, ContentSource source)
+        public string GetActualAssetKey(string key, ContentSource source = ContentSource.ModFolder)
         {
             switch (source)
             {
                 case ContentSource.GameContent:
-                    return this.ContentManager.NormaliseAssetName(this.StripXnbExtension(key));
+                    return this.ContentManager.NormaliseAssetName(key);
 
                 case ContentSource.ModFolder:
-                    string contentPath = Path.Combine(this.ModFolderPathFromContent, key);
-                    return this.ContentManager.NormaliseAssetName(this.StripXnbExtension(contentPath));
+                    FileInfo file = this.GetModFile(key);
+                    return this.ContentManager.NormaliseAssetName(this.GetModAssetPath(key, file.FullName));
 
                 default:
                     throw new NotSupportedException($"Unknown content source '{source}'.");
@@ -145,13 +142,29 @@ namespace StardewModdingAPI.Framework
                 throw new ArgumentException("The asset key or local path contains invalid characters.");
         }
 
-        /// <summary>Strip the .xnb extension from an asset key, since it's assumed by the underlying content manager.</summary>
-        /// <param name="key">The asset key.</param>
-        private string StripXnbExtension(string key)
+        /// <summary>Get a file from the mod folder.</summary>
+        /// <param name="path">The asset path relative to the mod folder.</param>
+        private FileInfo GetModFile(string path)
         {
-            if (key.EndsWith(".xnb", StringComparison.InvariantCultureIgnoreCase))
-                return key.Substring(0, key.Length - 4);
-            return key;
+            path = Path.Combine(this.ModFolderPath, this.ContentManager.NormalisePathSeparators(path));
+            FileInfo file = new FileInfo(path);
+            if (!file.Exists && file.Extension == "")
+                file = new FileInfo(Path.Combine(this.ModFolderPath, path + ".xnb"));
+            return file;
+        }
+
+        /// <summary>Get the asset path which loads a mod folder through a content manager.</summary>
+        /// <param name="localPath">The file path relative to the mod's folder.</param>
+        /// <param name="absolutePath">The absolute file path.</param>
+        private string GetModAssetPath(string localPath, string absolutePath)
+        {
+#if SMAPI_FOR_WINDOWS
+            // XNA doesn't allow absolute asset paths, so get a path relative to the content folder
+            return Path.Combine(this.ModFolderPathFromContent, localPath);
+#else
+            // MonoGame is weird about relative paths on Mac, but allows absolute paths
+            return absolutePath;
+#endif
         }
 
         /// <summary>Get a directory path relative to a given root.</summary>
@@ -181,14 +194,13 @@ namespace StardewModdingAPI.Framework
                 throw new NotSupportedException("Can't load a PNG file while the game is drawing to the screen. Make sure you load content outside the draw loop.");
 
             // process texture
+            SpriteBatch spriteBatch = Game1.spriteBatch;
+            GraphicsDevice gpu = Game1.graphics.GraphicsDevice;
             using (RenderTarget2D renderTarget = new RenderTarget2D(Game1.graphics.GraphicsDevice, texture.Width, texture.Height))
-            using (SpriteBatch spriteBatch = new SpriteBatch(Game1.graphics.GraphicsDevice))
             {
-                //Viewport originalViewport = Game1.graphics.GraphicsDevice.Viewport;
-
-                // create blank slate in render target
-                Game1.graphics.GraphicsDevice.SetRenderTarget(renderTarget);
-                Game1.graphics.GraphicsDevice.Clear(Color.Black);
+                // create blank render target to premultiply
+                gpu.SetRenderTarget(renderTarget);
+                gpu.Clear(Color.Black);
 
                 // multiply each color by the source alpha, and write just the color values into the final texture
                 spriteBatch.Begin(SpriteSortMode.Immediate, new BlendState
@@ -214,16 +226,17 @@ namespace StardewModdingAPI.Framework
                 spriteBatch.Draw(texture, texture.Bounds, Color.White);
                 spriteBatch.End();
 
-                // release the GPU
-                Game1.graphics.GraphicsDevice.SetRenderTarget(null);
-                //Game1.graphics.GraphicsDevice.Viewport = originalViewport;
+                // release GPU
+                gpu.SetRenderTarget(null);
 
-                // store data from render target because the RenderTarget2D is volatile
+                // extract premultiplied data
                 Color[] data = new Color[texture.Width * texture.Height];
                 renderTarget.GetData(data);
 
-                // unset texture from graphic device and set modified data back to it
-                Game1.graphics.GraphicsDevice.Textures[0] = null;
+                // unset texture from GPU to regain control
+                gpu.Textures[0] = null;
+
+                // update texture with premultiplied data
                 texture.SetData(data);
             }